# Swagger/src/main/webapp/swagger.yaml
openapi: 3.0.0
info:
  title: OpenDataAPI - UoC
  description: <h3>This page provides different formats for the student population data, available from the University of Crete Open Data Catalog.</h3><br/> <h4>Note that,  <u>graduation year K</u> refers to the minimum number of required year for graduation by each department.<h4>
  version: 1.0.0
servers:
  - url: 'http://147.52.17.140:8080/Swagger/'
...
paths:
...
  '/open/services/phd/getGrandTotalofAcademicYears/{year}':
    get:
      tags:
        - PHD
      summary: Get grand total of PHD students of all academic years per year
      parameters:
        - name: year
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            enum:
              - 2006-07
              - 2007-08
              - 2008-09
              - 2009-10
              - 2010-11
              - 2011-12
              - 2012-13
              - 2013-14
              - 2014-15
              - 2015-16
              - 2016-17
              - 2017-18
              - 2018-19
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
            text/html:
              schema:
                type: string
            text/plain:
              schema:
                type: string
            image/png:
              schema:
                type: string
  '/open/services/phd/getGrandTotalofGraduationYears/{year}':
    get:
      tags:
        - PHD
      summary: Get grand total of PHD students of all graduation years per year
      parameters:
        - name: year
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            enum:
              - 2006-07
              - 2007-08
              - 2008-09
              - 2009-10
              - 2010-11
              - 2011-12
              - 2012-13
              - 2013-14
              - 2014-15
              - 2015-16
              - 2016-17
              - 2017-18
              - 2018-19
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
            text/html:
              schema:
                type: string
            text/plain:
              schema:
                type: string
            image/png:
              schema:
                type: string
...
  '/open/services/grad/getGrandTotalofAcademicYears/{year}':
    get:
      tags:
        - Graduate
      summary: Get grand total of Graduate-students of all academic years per year
      parameters:
        - name: year
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            enum:
              - 2006-07
              - 2007-08
              - 2008-09
              - 2009-10
              - 2010-11
              - 2011-12
              - 2012-13
              - 2013-14
              - 2014-15
              - 2015-16
              - 2016-17
              - 2017-18
              - 2018-19
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
            text/html:
              schema:
                type: string
            text/plain:
              schema:
                type: string
            image/png:
              schema:
                type: string
  '/open/services/grad/getGrandTotalofGraduationYears/{year}':
    get:
      tags:
        - Graduate
      summary: Get grand total of Graduate-students of all graduation years per year
      parameters:
        - name: year
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            enum:
              - 2006-07
              - 2007-08
              - 2008-09
              - 2009-10
              - 2010-11
              - 2011-12
              - 2012-13
              - 2013-14
              - 2014-15
              - 2015-16
              - 2016-17
              - 2017-18
              - 2018-19
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
            text/html:
              schema:
                type: string
            text/plain:
              schema:
                type: string
            image/png:
              schema:
                type: string
...